Move PetSearch to services
