+++ app/src/main/java/com/codepath/apps/critterfinder/services/PetSearch.java
@@ -1,9 +1,10 @@
-package com.codepath.apps.critterfinder.models;
+package com.codepath.apps.critterfinder.services;
 
 import android.util.Log;
 import android.widget.ProgressBar;
 
 import com.codepath.apps.critterfinder.PetFinderHttpClient;
+import com.codepath.apps.critterfinder.models.PetModel;
 import com.loopj.android.http.JsonHttpResponseHandler;
 
 import org.json.JSONArray;
